Also accept null als value.
Sometimes you want to pass null as value (replace a template variable with an empty value), which you can't because it's not accepted, however it's pretty easy to cast null to empty string so added support for it. Side effect of this change is that PHP version 7.1 is required to use this feature, so from now on this is the minimum php version to use this package.